tvserver:switch audio using switch channel[1/1]
PD# SWPL-1158

Problem:
when switch audio, the video be stuck for a long time. it's sync issue.
but the schedule is hard, so using the temp solution.
when tsync work, please delete REPLAY_SWITCH_AUDIO_TRACK define
for removing this workarund.

Solution:
when switch audio,it will stop av, and update audio parameter and play.

Verify:
Verify livetv in P321 and timeshift in r321 using android p.

Change-Id: I71ebe3ffeeae3bacb0cffdd0199eb4f9d4adb777
Signed-off-by: Chuangcheng Peng <chuangcheng.peng@amlogic.com>